Use SDL_RWops for reading room files
Android makes accessing application's assets needlessly difficult,
and performing a simple open("directory/file.txt") does not work.
Instead, files have to be accessed via an AssetManager. Thankfully,
we don't have to write our own code for talking to the manager, as
SDL_RWFromFile() is able to handle that for us.

This code introduces a simple LineReader object type that serves
as a wrapper around SDL's RWops, allowing the Room code to focus
on fetching and parsing single lines of a file, leaving the actual
job of reading the file and splitting it into lines to the Reader.