Remove all references to new content
The ambitious plan was to introduce a whole new game world for v2.0,
along with some new enemies and other content. The reality is that
porting to SDL2 and adapting the game to Android proved difficult
enough. Should I decide to revisit the idea, it will probably be
for v3.0 (which is not as unlikely as it sounds, since development
of SDL3 has already been announced).