Explicitly set global vars during Startup()
Up until now, some global vars were set using the Initialization
clause of the units they were declared in. This is fine when developing
for desktop platforms, since we fully control the program's lifetime;
we can be sure that the main function is called exactly once. However,
on Android, the game is built as a dynamically-loaded library, and the
governing program can launch the main function multiple times. When this
happens, global vars retain their previous values. Explicitly setting
their value in the Startup() function ensures any previous values are
overwritten with the expected defaults.