Use custom resizable buffers for object lists
Until now, the game's entity management - handling of bullets, enemies,
and gibs - was based on Pascal's dynamic arrays feature. When a new
entity was added to one of the lists, the array was extended; upon
changing the active game room, an array was truncated to zero entries.
This effectively meant that the game was constantly reallocating memory
and moving data around, which is... definitely less than ideal.

This commit adds a resizable buffer object type to the game code.
Internally, the buffer still uses a dynamic array; however, it keeps
separate counters for its capacity and number of entries stored.
Thanks to this, memory will be reallocated only when the buffer's
capacity is exceeded; flushing the contents of the buffer and then
filling it up again will reuse the existing memory block.