Tint menu items in green/red pulse
This commit makes interactable menu items have a green/red tint, instead
of being rendered in plain white / grey. This should improve visibility
on mobile platforms, where it might have been a bit difficult to figure
out which parts of the menu were interactive, and which were just
informative (as they'd both be rendered in white).

The colours pulse slightly, to invite the user to interact with them
and make it even more apparent which elements are interactive.