Replace --compat-v1 config option with --compat
Having a --compat-v1 options suggests that maybe, possibly, if the game
were to have a v3.0 release, it would come with a --compat-v2 option.
Having separate options for each compat version would mean that the game
would have to support every possible combination of compat switches.

With a single option, compatibility can be thought of in terms of
levels; setting it to some level also includes all the higher ones.