Unify slide and thanks/stats screen event loops
The DisplaySlide and ShowOutro functions each had a separate event loop,
with slightly mismatching code. This was confusing from the player
perspective (slight different behaviour) and unwieldy from the developer
perspective. This commit reworks the slide code by introducing the
ShowSlides function, which takes an array of render functions and data
pointers, and then goes over said slides. This allows the individual
functions to focus solely on rendering, while leaving the event handling
to ShowSlides.

This change is applied to both the outro and the intro.