Remove the Children field from the Enemy type
This field was added back when I was planning for v2.0 of the game to
come with an entire new gameworld, with new enemies. That plan
eventually got scrapped, but the feature remained, unused. What's worse,
with the switch from dynamic arrays to custom resizable buffers
(commit 9cfeec26714009e7308e4d855f1bb0b437d29532), the code got broken
in a very subtle way.

This commit removes all code related to this field; should the idea
of a new gameworld resurface, this feature should be implemented anew.