Fix text colour in thanks/stats screen
The game's text rendering functions take an optional colour parameter;
when set to NIL, the text is rendered using the previously applied
colour setting. Because of a recent change that added green/red tints
to menu items (commit 5e410d59506a76ff90818d306df636f5c11ad430), and
because the rendering functions for the thanks & stats screen did not
set the text colour at any time, the text on said screens was rendered
with the menu's green tint, instead of white.