Display current keybinds in "bind keys" screen
The screen is very minimalistic - it shows only the name of the action
that the player should bind to a key. With each key provided, it goes
to the next action and that's it. This commit makes the interface more
helpful by changing it to display a list of all keybindings. The binding
logic is still the same - it goes over all possible actions,
sequentially.