Add Shared.GetTicks() function
The way time handling works in the game is that the Shared unit exposes
a function called GetDeltaTime(), which returns the number of ticks
(milliseconds) since the last time it was called. Under the hood, this
pretty much only calls SDL_GetTicks(), calculates the difference from
the previous value (stored as a static variable), stores the new value
and returns said difference.

Despite this, there is no way to directly retrieve the number of ticks,
so any place in the game that wants to perform some timer-based actions
has to manually keep track of time passed.

This commit adds a global function that can be used to retrieve the
current tick count. This function is then used to simplify rendering the
toast notification and menu colour "breathing" effect.