Fix hero's love missing from the starting room
Internally, the game treats the woman as a special type of colour
artefact; this allows to re-use some code. When entering a room which
contains an artefact, the game checks its state - whether it is
currently carried (STATE_PICKED) or has already been given to the woman
(STATE_GIVEN); this prevents a colour from appearing twice. The code
neglected to account for the fact that, even though she is (technically)
an artefact, the woman should be exempt from this check.

The bug has not manifested itself purely by dumb luck. The code would
read past the "colour state" array and use that value to determine
whether the woman should appear or not. This data is stored in a global
variable; the compiler lays those out in memory in the same order they
are declared in. So, what is located right past this array? Currently
active artefact's data - the first field of which is an "IsSet" boolean.
This was always set to false (which equals 0 which equals STATE_NONE)
whenever the game room was changed, right before the room's script
was executed.