Remember the order in which colours are given
This commit makes the game remember the order in which the player
brought the colour artefacts back to central area. This information
is also stored inside the save file. An error correction & order
guessing mechanism has also been implemented to allow the game to handle
saves where this info is missing or contains bogus values.