Rework "holding colour" outro slides
The outro begins with a narration on how the hero brought each colour
to his love. Unfortunately, all the slides use the same graphics,
with the hero holding some generic star item (which does not even
appear in the game); only the text is different.

This commit reworks the outro slides so that the hero is now presenting
the actual items that the player picks up. To better fit with the theme
of the game, the items retain their in-game colours, instead of
appearing pure-white, like the rest of the drawings.