Track which controller was last used
The game keeps a pointer to the "preferred" controller; this is used to
determine which controller to send rumble events to. This value is
updated only when a controller is (un)plugged.

This commit changes the behaviour so that the value is updated whenever
a user interacts with a controller (e.g. presses a button). This should
be more intuitive, as the game reacts better to the user's actions.