Show a toast regarding controllers after init
The game would show a toast if a controller was (un)plugged after the
game finished loading, but no information was provided to the user
regarding controllers that were already present when the game was
launched. This commit makes the game display the "controller found"
toast when it first finishes loading (if applicable, that is).