Store dead zone size as int+float pair
The game keeps the controller dead zone size as a signed 16-bit int,
since that's what the SDL API uses when reporting axis values. This
works fine when handling controller events; however, for showing the
value to the user, a percentage is more helpful. Because of rounding
errors involved in converting the value to a percentage and back, the
gamepad settings menu could get "stuck" - the user would try to change
the dead zone size, but end up with the same value.

This commit makes the game store the dead zone size as a pair of values;
a raw int value and a floating-point percentage. Both values can be read
and written to at any time; writing to one will cause the other to be
re-calculated.

The raw value is now used only when handling events. The config file
stores and reads the percentage, and that's also what's used when
tweaking the value in the menu.