Use SDL_GetTicks64() on 64-bit platforms
SDL2 provides two functions for timekeeping: the older SDL_GetTicks()
and the newer SDL_GetTicks64(). As the names suggest, the second one
returns a 64-bit value, whereas the first one returns a 32-bit value.

This commit makes the game base its timekeeping on the 64-bit function
if its built for a 64-bit target.