Fix invalid menu clicks when toast is shown
This commit fixes the menus mis-calculating which element was clicked
if the mouseclick arrived while a toast was being shown.

The click-handling code was mistakenly using the font's current scale,
instead of the scale set in the menu properties (which is applied
to the font whenever Menu.Draw() was called). This meant that things
would work fine if Menu.Draw() was the last thing to modify the font's
scale before the clicks were processed. Things would break if some
other part of the code changed the scale in the meantime - say, like
toast rendering does.