Add intermediate "room distance travelled" counter
This commit changes the way the "distance travelled" stat is calculated.
Instead of increasing the counter on each frame the hero moved,
an intermediate "room distance travelled" counter is introduced.
Each time the hero moves, it's this intermediate counter that is
increased; the main counter is increased only when changing between
rooms, or exiting from the game to the main menu (or victory screen).

This should help reduce potential floating-point rounding errors that
may have occurred when adding the miniscule per-frame movement values
to the main counter. It will also reduce the number of IsSet() checks
performed on the main counter.